requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
python-dotenv==1.0.1
solace-pubsubplus==1.8.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
